+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   from 'H' to 'depth' (issue #137).
 - Removed 'depth' attribute in optimise load functions, for it is already included in the Borefield object (issue #317).
 - Added 'depth' attribute to get_Rb and calculate_Rb functions in the Borehole class (issue #317).
+- Changed 'depth' attribute to 'length' in print_temperature_profile_fixed_length, calculate_temperatures (issue #317).
 
 ## [2.3.0] - 2024-11-05